requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.0
pytorch-grad-cam>=1.4.8
wandb>=0.16.0
tensorboard>=2.14.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
omegaconf>=2.3.0
tqdm>=4.66.0
Pillow>=10.0.0
numpy>=1.24.0
